test: replace renderer usage, combine files
